Working Alpha version with movements
