+++ project/game/__init__.py
@@ -1,3 +1,3 @@
 """
-The game package contains specific classes for playing Bat.
+The game package contains specific classes for playing the game.
 """
