+++ project/game/constants.py
@@ -1,4 +1,3 @@
-
 #Screen Constants
 SCREEN_WIDTH = 800
 SCREEN_HEIGHT = 600
@@ -7,3 +6,6 @@ SCREEN_TITLE = "TEAM 01 ARCADE GAME"
 #Sprite Constants scaled in relation to the original pics
 CHARACTER_SCALING = 1
 TILE_SCALING = 0.5
+
+# Movement speed
+PLAYER_MOVEMENT_SPEED = 5
