+++ project/game/gems.py
@@ -4,7 +4,3 @@ class Player(arcade.Sprite):
 
     def __init__(self, path, scale):
         super().__init__(path, scale) 
-    
-
-        
-
